+++ .devcontainer/dev_launcher
@@ -4,4 +4,4 @@
 # is considered:
 cd /workspace/.devcontainer
 
-drs3
+dcs
