+++ Dockerfile
@@ -29,4 +29,4 @@ USER appuser
 
 ENV PYTHONUNBUFFERED=1
 
-ENTRYPOINT ["drs3"]
+ENTRYPOINT ["dcs"]
