+++ README.md
@@ -1,6 +1,6 @@
-![tests](https://github.com/ghga-de/drs3/actions/workflows/unit_and_int_tests.yaml/badge.svg)
-[![codecov](https://codecov.io/gh/ghga-de/drs3/branch/main/graph/badge.svg?token=GYH99Y71CK)](https://codecov.io/gh/ghga-de/drs3)
-# DRS3 - a GA4GH drs compliant service for delivering files from S3
+![tests](https://github.com/ghga-de/dcs/actions/workflows/unit_and_int_tests.yaml/badge.svg)
+[![codecov](https://codecov.io/gh/ghga-de/dcs/branch/main/graph/badge.svg?token=GYH99Y71CK)](https://codecov.io/gh/ghga-de/dcs)
+# Download Controller Service - a GA4GH drs compliant service for delivering files from S3
 
 This service implements the GA4GH DRS, while providing the option to serve files from localstack S3.
 
@@ -12,16 +12,16 @@ An extensive documentation can be found [here](...) (coming soon).
 ### Installation
 We recommend using the provided Docker container.
 
-A pre-build version is available at [docker hub](https://hub.docker.com/repository/docker/ghga/drs3):
+A pre-build version is available at [docker hub](https://hub.docker.com/repository/docker/ghga/dcs):
 ```bash
 # Please feel free to choose the version as needed:
-docker pull ghga/drs3:<version>
+docker pull ghga/dcs:<version>
 ```
 
 Or you can build the container yourself from the [`./Dockerfile`](./Dockerfile):
 ```bash
 # Execute in the repo's root dir:
-docker build -t ghga/drs3:<version> .
+docker build -t ghga/dcs:<version> .
 ```
 
 For production-ready deployment, we recommend using Kubernetes, however,
@@ -29,7 +29,7 @@ for simple use cases, you could execute the service using docker
 on a single server:
 ```bash
 # The entrypoint is preconfigured:
-docker run -p 8080:8080 ghga/drs3:<version>
+docker run -p 8080:8080 ghga/dcs:<version>
 ```
 
 If you prefer not to use containers, you may install the service from source:
@@ -38,14 +38,14 @@ If you prefer not to use containers, you may install the service from source:
 pip install .
 
 # to run the service:
-drs3
+dcs
 ```
 
 ### Configuration:
 The [`./example-config.yaml`](./example-config.yaml) gives an overview of the available configuration options.
-Please adapt it, rename it to `.drs3.yaml`, and place it to one of the following locations:
-- in the current working directory were you are execute the service (on unix: `./.drs3.yaml`)
-- in your home directory (on unix: `~/.drs3.yaml`)
+Please adapt it, rename it to `.dcs.yaml`, and place it to one of the following locations:
+- in the current working directory were you are execute the service (on unix: `./.dcs.yaml`)
+- in your home directory (on unix: `~/.dcs.yaml`)
 
 The config yaml will be automatically parsed by the service.
 
@@ -54,8 +54,8 @@ The config yaml will be automatically parsed by the service.
 All parameters mentioned in the [`./example-config.yaml`](./example-config.yaml)
 could also be set using environment variables or file secrets.
 
-For naming the environment variables, just prefix the parameter name with `drs3_`,
-e.g. for the `host` set an environment variable named `drs3_host`
+For naming the environment variables, just prefix the parameter name with `dcs_`,
+e.g. for the `host` set an environment variable named `dcs_host`
 (you may use both upper or lower cases, however, it is standard to define all env
 variables in upper cases).
