+++ dcs/config.py
@@ -26,7 +26,7 @@ from ghga_service_chassis_lib.s3 import S3ConfigBase
 LogLevel = Literal["critical", "error", "warning", "info", "debug", "trace"]
 
 
-@config_from_yaml(prefix="drs3")
+@config_from_yaml(prefix="dcs")
 class Config(ApiConfigBase, PubSubConfigBase, PostgresqlConfigBase, S3ConfigBase):
     """Config parameters and their defaults."""
 
@@ -37,12 +37,12 @@ class Config(ApiConfigBase, PubSubConfigBase, PostgresqlConfigBase, S3ConfigBase
     #   - the S3 interface
     # are inherited.
 
-    # Following config parameter are specifically needed for drs3:
+    # Following config parameter are specifically needed for dcs:
     api_route: str = "/ga4gh/drs/v1"
     drs_self_url: str = "drs://localhost:8080/"
     custom_spec_url: Optional[str] = None
 
-    service_name: str = "drs3"
+    service_name: str = "dcs"
     topic_name_stage_request: str = "non_staged_file_requested"
     topic_name_file_staged: str = "file_staged_for_download"
     topic_name_file_registered: str = "file-internally-registered"
