+++ setup.cfg
@@ -14,10 +14,10 @@
 # limitations under the License.
 
 [metadata]
-name = drs3
-version = attr: drs3.__version__
-description = drs3 - a GA4GH drs compliant service for delivering files from S3
-url = https://github.com/ghga-de/drs3
+name = dcs
+version = attr: dcs.__version__
+description = dcs - a GA4GH drs compliant service for delivering files from S3
+url = https://github.com/ghga-de/dcs
 long_description = file: README.md
 long_description_content_type = text/markdown; charset=UTF-8
 author = German Human Genome Phenome Archive (GHGA)
@@ -49,7 +49,7 @@ python_requires = >= 3.9
 
 [options.entry_points]
 console_scripts =
-    drs3 = drs3.__main__:run
+    dcs = dcs.__main__:run
 
 [options.extras_require]
 dev =
